refactor: move common methods from project to common
