+++ common/url_test.go
@@ -1,4 +1,4 @@
-package project
+package common
 
 import (
 	"testing"
@@ -16,89 +16,89 @@ func TestProjectUrlJoinManifest(t *testing.T) {
 
 	///////////////////
 	base = "https://example.com/foo.git"
-	u, err = urlJoin(base, ".", "bar.git")
+	u, err = URLJoin(base, ".", "bar.git")
 	assert.Nil(err)
 	assert.Equal("https://example.com/bar.git", u)
 
-	u, err = urlJoin(base, "..", "bar.git")
+	u, err = URLJoin(base, "..", "bar.git")
 	assert.Nil(err)
 	assert.Equal("https://example.com/bar.git", u)
 
 	///////////////////
 	base = "https://github.com/jiangxin/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/my/repo", u)
 
-	u, err = urlJoin(base, "../..", "my/repo")
+	u, err = URLJoin(base, "../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/my/repo", u)
 
 	///////////////////
 	base = "https://github.com/jiangxin/manifest/"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/my/repo", u)
 
-	u, err = urlJoin(base, "../..", "my/repo")
+	u, err = URLJoin(base, "../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/my/repo", u)
 
 	///////////////////
 	base = "https://github.com/jiangxin/manifest.git/"
 
-	u, err = urlJoin(base, ".", "my/repo")
+	u, err = URLJoin(base, ".", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/jiangxin/my/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://github.com/my/repo", u)
 
 	///////////////////
 	base = "ssh://git@github.com/jiangxin/manifest.git/"
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@github.com/my/repo", u)
 
 	///////////////////
 	base = "file:///root/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("file:///root/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("file:///my/repo", u)
 
-	u, err = urlJoin(base, "../../..", "my/repo")
+	u, err = URLJoin(base, "../../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("file:///my/repo", u)
 
 	///////////////////
 	base = "git@github.com:jiangxin/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@github.com/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@github.com/my/repo", u)
 
-	u, err = urlJoin(base, "../../..", "my/repo")
+	u, err = URLJoin(base, "../../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@github.com/my/repo", u)
 }
@@ -114,7 +114,7 @@ func TestProjectUrlJoinAbs(t *testing.T) {
 	///////////////////
 	base = "https://github.com/jiangxin/manifest.git"
 
-	u, err = urlJoin(base, "https://example.com/projects", "repo")
+	u, err = URLJoin(base, "https://example.com/projects", "repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/projects/repo", u)
 
@@ -131,134 +131,82 @@ func TestProjectUrlJoinManifestWithSpace(t *testing.T) {
 	///////////////////
 	base = "https://example.com/repo dir/jiangxin/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/my/repo", u)
 
-	u, err = urlJoin(base, "../..", "my/repo")
+	u, err = URLJoin(base, "../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/my/repo", u)
 
 	///////////////////
 	base = "https://example.com/repo dir/jiangxin/manifest/"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/my/repo", u)
 
-	u, err = urlJoin(base, "../..", "my/repo")
+	u, err = URLJoin(base, "../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/my/repo", u)
 
 	///////////////////
 	base = "https://example.com/repo dir/jiangxin/manifest.git/"
 
-	u, err = urlJoin(base, ".", "my/repo")
+	u, err = URLJoin(base, ".", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/jiangxin/my/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("https://example.com/repo dir/my/repo", u)
 
 	///////////////////
 	base = "ssh://git@example.com/repo dir/jiangxin/manifest.git/"
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@example.com/repo dir/my/repo", u)
 
 	///////////////////
 	base = "file:///root/repo dir/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("file:///root/repo dir/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("file:///root/my/repo", u)
 
-	u, err = urlJoin(base, "../../..", "my/repo")
+	u, err = URLJoin(base, "../../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("file:///my/repo", u)
 
-	u, err = urlJoin(base, "../../../..", "my/repo")
+	u, err = URLJoin(base, "../../../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("file:///my/repo", u)
 
 	///////////////////
 	base = "git@example.com:repo dir/jiangxin/manifest.git"
 
-	u, err = urlJoin(base, ".", "repo")
+	u, err = URLJoin(base, ".", "repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@example.com/repo dir/jiangxin/repo", u)
 
-	u, err = urlJoin(base, "..", "my/repo")
+	u, err = URLJoin(base, "..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@example.com/repo dir/my/repo", u)
 
-	u, err = urlJoin(base, "../../..", "my/repo")
+	u, err = URLJoin(base, "../../..", "my/repo")
 	assert.Nil(err)
 	assert.Equal("ssh://git@example.com/my/repo", u)
-}
-
-func TestMatchGroups(t *testing.T) {
-	var (
-		match, groups string
-	)
-
-	assert := assert.New(t)
-
-	match = ""
-	groups = ""
-	assert.True(MatchGroups(match, groups))
-
-	match = "default"
-	groups = ""
-	assert.True(MatchGroups(match, groups))
-
-	match = "all"
-	groups = ""
-	assert.True(MatchGroups(match, groups))
-
-	match = "g3"
-	groups = "g1,g2"
-	assert.False(MatchGroups(match, groups))
-
-	match = "-g1,g2"
-	groups = "g1,g2"
-	assert.True(MatchGroups(match, groups))
-
-	match = "g1,-g2"
-	groups = "g1,g2"
-	assert.False(MatchGroups(match, groups))
-
-	match = ""
-	groups = "g1,notdefault"
-	assert.False(MatchGroups(match, groups))
-
-	match = "default"
-	groups = "g1,notdefault"
-	assert.False(MatchGroups(match, groups))
-
-	match = "all"
-	groups = "g1,notdefault"
-	assert.True(MatchGroups(match, groups))
-
-	match = "g1"
-	groups = "g1,notdefault"
-	assert.True(MatchGroups(match, groups))
-
-	match = "g2"
-	groups = "g1,notdefault"
-	assert.False(MatchGroups(match, groups))
 }
